chore: final project polish, update README and .gitignore

+++ backend/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, UniqueConstraint, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from database import Base
@@ -36,8 +36,8 @@ class Product(Base):
     sku = Column(String, nullable=False)
     description = Column(String, nullable=True)
     quantity_on_hand = Column(Integer, default=0)
-    cost_price = Column(Float, nullable=True)
-    selling_price = Column(Float, nullable=True)
+    cost_price = Column(Numeric(10, 2), nullable=True)
+    selling_price = Column(Numeric(10, 2), nullable=True)
     low_stock_threshold = Column(Integer, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
